Hybrid_Search_RAG_App/requirements.txt
python-dotenv==1.0.1
pinecone-client==5.0.1
pinecone-text==0.9.0
langchain-community==0.3.17
fastembed==0.5.1
langchain-core==0.3.34
docling==2.21.0
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
huggingface-hub==0.28.1
nltk==3.9.1
torch==2.6.0
streamlit==1.42.0
